[#408] Move OriginatingClasses annotation to protostream-core
