+++ core/src/main/java/org/infinispan/protostream/annotations/impl/OriginatingClasses.java
@@ -1,4 +1,4 @@
-package org.infinispan.protostream.annotations.impl.processor;
+package org.infinispan.protostream.annotations.impl;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
